requirements.txt
bertopic
einops
hdbscan
ipykernel
matplotlib
numpy
pandas
plotly
pyspark
pyarrow
sentence-transformers
scikit-learn
tokenizers
torch
transformers
umap-learn




